Moved TRD from SBP repo to TRD repo
The document WordPress on LAMP on SUSE Linux Enterprise Server has been
moved from the SBP repo to the TRD repo. This was needed to be able to
build from the TRD configuration for DocServ the new TRD docs under the docset linux in category getting
started from one builddocs source which should be the TRD repo.