Test: changed includes to see if document builds
We got an error message that the include documents are missing. They are
there in the same directory as the main content adoc, but as symlinks.
These symlinks link to another but higher level subdirectory named
common. Changed now the path in the main content adoc file to see if the
symlinked documents are found now.