Updated document to new TRD layout
DC file and docinfo file have been updated to reflect the new TRD
layout. Issues: the docdate showed up twice, had to outcomment it in
docinfo.xml to fix the issue. The SUSE logo needs to be adjusted to
suse.svg via reference or symlink I assume. I had to outcomment it for
now because it created layout problems.